requirements.txt
jsonpickle
appdirs
contracts
tinydb
selenium
guessit
lxml
requests
pysynthetic
virtualenv
watchdog